move operator classes around to different files
