+++ src/Makefile.am
@@ -3,7 +3,7 @@ LEXERSOURCES = njnr.tab.cpp lex.yy.cc location.hh njnr.tab.hpp
 
 njnrSOURCES = command.cpp compiler_core.cpp debug.cpp
 njnrSOURCES += getopt.cpp helper.cpp njnrLexer.cpp List.cpp ListNode.cpp
-njnrSOURCES += main.cpp operator.cpp parser_functions.cpp
+njnrSOURCES += main.cpp Operator.cpp parser_functions.cpp
 njnrSOURCES += CodeGenerator.cpp common/symbol_table.cpp common/symbol_table_stack.cpp
 njnrSOURCES += common/symbol_table_stackX.cpp common/symbol_table_entry.cpp
 njnrSOURCES += expressions/expressions.cpp
@@ -16,7 +16,7 @@ njnrSOURCES += include/ListNode.hpp include/Constant.hpp include/common/symbol_t
 njnrSOURCES += include/common/symbol_table_stack.hpp include/common/symbol_table.hpp
 njnrSOURCES += include/Compiler.hpp include/CodeGenerator.hpp include/common/symbol_table_entry.hpp
 njnrSOURCES += include/njnrLexer.hpp include/data.hpp include/helper.hpp
-njnrSOURCES += include/operator.hpp include/common/symbol_table_templates.h
+njnrSOURCES += include/Operator.hpp include/common/symbol_table_templates.h
 njnrSOURCES += include/common/symbol_table_stack_templates.h include/ReturnPacket.hpp
 njnrSOURCES += include/BaseData.hpp include/ParameterBinding.hpp include/Statement.hpp
 njnrSOURCES += include/FunctionBinding.hpp include/Identifier.hpp include/VariableBinding.hpp
@@ -24,6 +24,7 @@ njnrSOURCES += include/TSType.hpp include/TSAbstractType.hpp include/Translation
 njnrSOURCES += include/FunctionHeader.hpp include/BasicListNode.hpp include/StatementListNode.hpp
 njnrSOURCES += include/TypeListNode.hpp include/IdentifierListNode.hpp include/TranslationUnitListNode.hpp
 njnrSOURCES += include/ParameterListNode.hpp include/ReturnPacketListNode.hpp
+njnrSOURCES += include/TSOperatorType.hpp
 njnr_SOURCES = $(LEXERSOURCES) $(njnrSOURCES)
 AUTOMAKE_OPTIONS = foreign
